requirements.txt
flask==2.3.3
boto3==1.28.29
python-dotenv==1.0.0
qrcode==7.4.2
pillow==11.0.0
Flask-WTF==1.1.1
pytest==7.4.0
pytest-flask==1.2.0
